requirements.txt
factory
faker
aniso8601==7.0.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.6.3
cairocffi==1.0.2
CairoSVG==2.4.0
casbin==0.8.4
certifi==2018.8.24
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cssselect2==0.2.1
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.0.2
django-allauth==0.38.0
django-braces==1.13.0
django-cleanup==2.1.0
django-cors-headers==3.3.0
django-debug-toolbar==1.11
django-environ==0.4.5
django-filter==2.1.0
django-graphql-jwt==0.3.1
django-import-export==1.2.0
django-jalali==4.0.0
django-pdfkit==0.3.1
django-unused==0.1.4
django-webpack-loader==0.6.0
djangorestframework==3.9.0
environ==1.0
et-xmlfile==1.0.1
flake8==3.6.0
graphene==2.1.8
graphene-django==2.8.1
graphql-core==2.3.1
graphql-relay==2.0.1
html5lib==1.0b8
idna==2.7
isort==4.3.21
jdatetime==3.0.2
jdcal==1.4
json-tricks==3.12.2
lazy-object-proxy==1.4.3
Markdown==3.0.1
mccabe==0.6.1
more-itertools==7.0.0
mysqlclient==2.0.1
nested-dict==1.61
oauthlib==2.1.0
odfpy==1.4.0
openpyxl==2.5.12
pdfkit==0.6.0
Pillow==5.2.0
pluggy==0.9.0
promise==2.3
py==1.8.0
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
PyJWT==1.7.1
pylint==2.4.4
PyPDF2==1.26.0
Pyphen==0.9.5
pytesseract==0.2.4
pytest==4.4.1
pytest-django==3.4.8
python-datauri==0.2.8
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
reportlab==3.5.23
requests==2.19.1
requests-oauthlib==1.0.0
Rx==1.6.1
selenium==3.14.0
simpleeval==0.9.10
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.0
tablib==0.12.1
tinycss2==1.0.2
typed-ast==1.4.1
unicodecsv==0.14.1
urllib3==1.23
virtualenv==16.0.0
WeasyPrint==47
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.3
xlrd==1.2.0
xlwt==1.3.0
uWSGI>=2.0.18,<2.1
pandas==1.0.3
pandas-datareader==0.8.1